# server/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"data": {}
				},
				"required": [
					"message",
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"LoginUser": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"RegisterUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"passwordAgain": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"passwordAgain"
				],
				"type": "object",
				"additionalProperties": true
			},
			"NewType": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": true
			},
			"NewProject": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"description",
					"type_id"
				],
				"type": "object",
				"additionalProperties": true
			},
			"NewTask": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"developer_id": {
						"type": "number",
						"format": "double"
					},
					"manager_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"description",
					"developer_id",
					"manager_id"
				],
				"type": "object",
				"additionalProperties": true
			},
			"NewDeveloper": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "redmine-server",
		"version": "Beadando#3",
		"description": "Redmine - 3. Csapat",
		"license": {
			"name": "GPL-3.0"
		},
		"contact": {
			"name": "Regenye Tamás | Kukk Péter | Szűcs Gergely ",
			"url": "https://github.com/KukkPeter/Redmine"
		}
	},
	"paths": {
		"/user/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Bejelentkezés során megkell adni egy email címet ami regisztrálva van a rendszerben és egy hozzá tartozó jelszót.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginUser"
							}
						}
					}
				}
			}
		},
		"/user/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Regisztráció során megkell adni egy email címet ami még nincs regisztrálva a rendszerben, egy nevet illetve egy jelszót és a hozzá tartozó jelszó megerősítőt.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterUser"
							}
						}
					}
				}
			}
		},
		"/user/logout": {
			"post": {
				"operationId": "LogoutUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Kijelentkezés a rendszerből.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/user/myself": {
			"get": {
				"operationId": "GetMyself",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "A headerben megadott JsonWebToken-ből vissza adja a felhasználót.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/user/{userId}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "A megadott userId alapján kikeresi a rendszerben tárolt felhasználót, feltéve hogy létezik ilyen ID-val felhasználó.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/types": {
			"get": {
				"operationId": "GetProjectTypes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Visszaadja a rendszerben tárolt projekt típusokat.",
				"tags": [
					"Types"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/types/{typeId}": {
			"get": {
				"operationId": "GetProjectTypeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Visszaadja a megadott typeId alapján a rendszerben tárolt projekt típust, feltéve hogy létezik a megadott ID-val projekt típus.",
				"tags": [
					"Types"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "typeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteTypeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Projekt típust töröl a rendszerből. Adminisztrátori jogosultság szükséges!",
				"tags": [
					"Types"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "typeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/types/new": {
			"post": {
				"operationId": "NewProjectType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Új projekt típust ad hozzá a rendszerhez. Adminisztrátori jogosultság szükséges!",
				"tags": [
					"Types"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewType"
							}
						}
					}
				}
			}
		},
		"/projects": {
			"get": {
				"operationId": "GetProjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Visszaadja a rendszerben tárolt projekteket.",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/projects/new": {
			"post": {
				"operationId": "NewProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Új projektet ad hozzá a rendszerhez. Adminisztrátori jogosultság szükséges!",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewProject"
							}
						}
					}
				}
			}
		},
		"/projects/{projectId}": {
			"get": {
				"operationId": "GetProjectByID",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Visszaadja a megadott projectId alapján a rendszerben tárolt projektet, feltéve hogy létezik a megadott ID-val projekt.",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteProjectByID",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Projekt törlése a rendszerből. Adminisztrátori jogosultság szükséges!",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/projects/{projectId}/tasks": {
			"get": {
				"operationId": "GetTasksForProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Visszaadja a megadott projectId alapján a rendszerben tárolt projekthez tartozó feladatokat, feltéve hogy létezik a megadott ID-val projekt.",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/projects/{projectId}/developers": {
			"get": {
				"operationId": "GetDevelopersForProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Visszaadja a megadott projectId alapján a rendszerben tárolt projekthez tartozó fejlesztőket, feltéve hogy létezik a megadott ID-val projekt.",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/projects/{projectId}/newTask": {
			"post": {
				"operationId": "NewTaskForProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "A projectId alapján új feladatot ad hozzá a projekthez, feltéve hogy létezik projekt a projectId-ban megadott ID-val.",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewTask"
							}
						}
					}
				}
			}
		},
		"/projects/deleteTask/{taskId}": {
			"delete": {
				"operationId": "DeleteTaskById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Feladat törlése a rendszerből. Adminisztrátori jogosultság szükséges!",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/developers": {
			"get": {
				"operationId": "GetDevelopers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Visszaadja a rendszerben tárolt fejlesztőket.",
				"tags": [
					"Developers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/developers/{developerId}": {
			"get": {
				"operationId": "GetDeveloperById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Visszaadja a megadott developerId alapján a rendszerben tárolt fejlesztőt, feltéve hogy létezik a megadott ID-val fejlesztő.",
				"tags": [
					"Developers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "developerId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteDeveloperById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Fejlesztőt töröl a rendszerből. Adminisztrátori jogosultság szükséges!",
				"tags": [
					"Developers"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "developerId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/developers/new": {
			"post": {
				"operationId": "AddNewDeveloper",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Új fejlesztőt ad hozzá a rendszerhez. Adminisztrátori jogosultság szükséges!",
				"tags": [
					"Developers"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewDeveloper"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8000/"
		}
	]
}